backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pydantic==2.7.4
pydantic-settings==2.3.3
python-dotenv==1.0.1
pymongo==4.7.3
pydantic[email]==2.7.4
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
